fix: bind replay resume digest to execution plan
