+++ test/skillgym/suites/agent-device-smoke-suite.ts
@@ -2493,11 +2493,20 @@ Treat the recovery message as a warning, not a fatal error. Use the exposed Sear
     contract: [
       'Replay path: ./replays/catalog-checkout.ad',
       'Selectors drifted after a UI label change',
-      'Goal: maintain the replay script in place',
+      '--update is accepted but retired: it never rewrites the replay file and reports healed: 0',
+      'A failed replay returns a divergence report with a suggested selector plus resume.from and resume.planDigest',
+      'Goal: repair the script explicitly, then resume from the failing step without rerunning earlier steps',
     ],
-    task: 'Plan the command to maintain the existing replay script after selector drift.',
-    outputs: [plannedCommand('replay'), /-u|--update/i, /\.\/replays\/catalog-checkout\.ad/i],
-    forbiddenOutputs: [/sed\s+-i/i, /open .*\.ad/i],
+    task: 'Plan the safe replay-maintenance workflow after selector drift.',
+    outputs: [
+      plannedCommand('replay'),
+      /\.\/replays\/catalog-checkout\.ad/i,
+      /--update.*(?:no-op|does not rewrite|retired)|(?:no-op|does not rewrite|retired).*--update/i,
+      /manual(?:ly)? (?:edit|repair)/i,
+      /--from\s+(?:<[^>]+>|\d+)/i,
+      /--plan-digest\s+(?:<[^>]+>|[a-f0-9]{8,})/i,
+    ],
+    forbiddenOutputs: [/sed\s+-i/i],
   }),
   makeCase({
     id: 'replay-maestro-compatibility-flow',
